requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
fastapi==0.112.2
FastAPI-SQLAlchemy==0.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.8
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
numpy==2.1.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg==3.2.1
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
rdkit==2024.3.5
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
